Some fixes

+++ src/main/java/es/uvlive/controllers/response/MessageResponse.java
@@ -9,7 +9,7 @@ public class MessageResponse extends BaseResponse {
     
 	private int idMessage;
 	private String text;
-	private int timestamp;
+	private long timestamp;
 	private String owner;
 
 	public int getIdMessage() {
@@ -28,11 +28,11 @@ public class MessageResponse extends BaseResponse {
 		this.text = text;
 	}
 
-	public int getTimestamp() {
+	public long getTimestamp() {
 		return timestamp;
 	}
 
-	public void setTimestamp(int timestamp) {
+	public void setTimestamp(long timestamp) {
 		this.timestamp = timestamp;
 	}
